backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
awsiotsdk==1.21.0
boto3==1.34.0
